// lib/type.ts
export enum BizCode {
  MARK_DROP = "MARK_DROP",
}

export enum RemoveType {
  AUTO = "AUTO",
  MANUAL = "MANUAL",
}

export enum TaskState {
  QUEUE = "QUEUE",
  PROCESSING = "PROCESSING",
  SUCCESS = "SUCCESS",
  FAIL = "FAIL",
  CANCEL = "CANCEL",
}


export enum PointChangeTypeEnum {
  RECHARGE = "RECHARGE",
  BUSINESS_CONSUME = "BUSINESS_CONSUME",
  FAILURE_RETURN = "FAILURE_RETURN",
}